ARROW-5150: [Ruby] Add Arrow::Table#raw_records
I'd like to add `raw_records` method in `Arrow::Table` class.
The same method of `RecordBatch` was added in #3587.

Author: Kenta Murata <mrkn@mrkn.jp>
Author: Kouhei Sutou <kou@clear-code.com>

Closes #4133 from mrkn/ruby_table_raw_records and squashes the following commits:

8d6b0d9c <Kouhei Sutou> Remove needless TODO
cc173459 <Kouhei Sutou> Unify tests for RecordBatch and Table
93967559 <Kenta Murata> Replace + with ++ for shifting row index
52cb4025 <Kenta Murata> Fix tags for errors
f7b88d4d <Kenta Murata> Add test cases for multiple columns and arrays
9ba352f7 <Kenta Murata> Add tests for StructArray
c47b5b34 <Kenta Murata> Add tests for ListArray
f6551fdc <Kenta Murata> Add tests for SparseUnionArray
a34c3603 <Kenta Murata> Add tests for DenseUnionArray
e8fb6c9b <Kenta Murata> Add Arrow::Table#raw_records
1e4f48ba <Kenta Murata> Rename
