+++ ruby/red-arrow/ext/arrow/arrow.cpp
@@ -34,6 +34,10 @@ extern "C" void Init_arrow() {
   rb_define_method(cArrowRecordBatch, "raw_records",
                    reinterpret_cast<rb::RawMethod>(red_arrow::record_batch_raw_records),
                    0);
+  auto cArrowTable = rb_const_get_at(mArrow, rb_intern("Table"));
+  rb_define_method(cArrowTable, "raw_records",
+                   reinterpret_cast<rb::RawMethod>(red_arrow::table_raw_records),
+                   0);
 
   red_arrow::cDate = rb_const_get(rb_cObject, rb_intern("Date"));
